requirements.txt
Flask
openai
python-dotenv
azure-core
azure-search-documents
gunicorn
